+++ emp-ot/ideal.h
@@ -7,7 +7,7 @@ template<typename T>
 class OTIdeal: public COT<T> { public:
 	using COT<T>::io;
 	using COT<T>::Delta;
-	size_t cnt = 0;
+	int64_t cnt = 0;
 	PRG prg;
 	OTIdeal(T * io, bool * delta = nullptr) {
 		this->io = io;
@@ -16,12 +16,12 @@ class OTIdeal: public COT<T> { public:
 			Delta = bool_to_block(delta);
 	}
 
-	void send_cot(block* data, size_t length) override {
+	void send_cot(block* data, int64_t length) override {
 		cnt+=length;
 		prg.random_block(data);
 	}
 
-	void recv_cot(block* data, const bool* b, size_t length) override {
+	void recv_cot(block* data, const bool* b, int64_t length) override {
 		cnt+=length;
 		prg.random_block(data);
 		for(int i = 0; i < length; ++i)
